feat: added environment_setup.py, a file in which the main class is defined for environment setup. empty_main.py gives an example of how to use this class.
